requirements.txt
matplotlib==3.6.2
networkx==2.8.8
Pillow==9.3.0
pytesseract==0.3.10
